Fix Public Access bug and enable the field by default
While reviewing issue #19 we realised that the PDF public_access logic wasn't configured correctly. This was correctly fixed using a filter and the new function was unit tested. Along with this, the public_access field was enabled by default (not by a filter) and adequate warnings were included for the user.

Resolves #19.