Allow Show/Hide Conditional Logic for PDFs
We had to abstract out the GFCommon::evaluate_conditional_logic() method so we can easily support Show AND Hide behaviour. This is because Gravity Forms returns true if the condition is met for both show and hide which was problematic. We standadised this in Helper_Misc so `true` is returned if the PDF should be shown and `false` is returned if not.