Fix header/footer manual display issue
This issue was caused because we were writing our core styles (which includes header/footer details) to the PDF before our template. This caused the software to create a new page (with no header/footer). The solution was to write the styles at the same time.

This may create a regression with v3 templates and we'll need to get a close eye for any issues.

Fixes #206

Tweak unit test so new changes pass