Create markup closer to our v3 style when using depreciated methods
Our premium v3 templates accepts the returned markup as an array and parses it to create columns. The way we had it set up in v4 was just to return the entire string.

To get around this issue we moved a stripped down version of the field output to depreciated.php.

Fixes #227

Add additional backwards compatible tweaks to ensure a smoother transition for more users