Preliminary fix for Tier 2 add-on.
There was a major issue with the way we were parsing the legacy Tier 2 add-on. It appeared to work correctly when only one Tier 2 template was being run but once you started using multiples it fell to pieces.

This patch reimagines how the legacy Tier 2 compatibility should function in Gravity PDF 4.0. Instead of handling it inside our Helper_PDF class, we are instead bypassing that functionality and doing the checks before hand in our Model_PDF and View_PDF functions (one for saving and one for viewing).

We also fixed a clean-up issue when resending PDFs via the admin area.

This patch still needs both unit testing and more hands on testing before merging.

Resolves #289

Add unit tests and optimise code