Fix legacy multisite template folder issue
Older versions of multisite installations didn't use a standardised upload path which messes with our folder structure. We'll switch to the initial primary blog (which did use the /wp-content/uploads/ directory) and then restore the current blog at the end.

Note: because BLOG_ID_CURRENT_SITE can be changed and the 'upload_path' option doesn't change with it we opted to hardcode the ID instead.

Resolves #271