Fix issue with Appearance tab not showing up on template change
This problem occurs when a template is saved with the Advanced Template option enabled. Then, when you try change the template to another the Appearance tab doesn't show up because the Advanced Template option is still activated. To resolve this I reset that option to No on template change. It's a legacy option anyway so I'm not concerned about doing it this way.

Fixes #333