Fix font caching bug and duplicate custom font types
mPDF caches the processed font files and saves it for later (see https://github.com/GravityPDF/gravity-pdf/pull/193). The problem occurs when a user drops a font file directly in the /fonts/ directory without installing it via our font installer (we're most effected by this when doing PDF integrations), loads some PDFs but then decides to install the font via the installer with the same name. Because of the font cache this causes problems. To get around this issue we are no clearing the fontdata cache when fonts are installed, edited or deleted.

While reviewing this problem we found that font files assigned to a group could still be included in mPDFs font array because we were only checking the top array key. To fix this problem we now recursively look through the multidimensional array and check if the font already exists.

Add unit test for our Helper_Misc::is_array() method