Apply HTML escaping during value() method
Instead of applying esc_html() during the html() method, we're going to apply it at the value() method so the $form_data array is appropriately protected as well.
This does break backwards compatibility and make it more difficult for users to apply conditional logic in custom templates (as there's encoded characters to consider - < > ' " &) but we'll note this in the developer documentation and upgrade guide.

Resolves #385