Allow Radio and Checkbox values to show HTML in PDFs
This is because these fields are set by the site owner and not the end user so it's presumed to be trusted. The exception to this is user-defined values in Radio fields which allows the "other" option.

With that said, all HTML is run through wp_kses_post() to prevent any nasty surprises.

Resolves #415