Redirect to our update screen when the user updates to a new major or minor release
Previously we restricted this to only show when updating to a minor release (4.x), but many missed out because we pushed our patches fairly quickly. That meant that even though they updated to a new minor release they didn't see the update message because they jumped straight to 4.x.x.

This patch resolves this issue by comparing the old and the new version numbers and seeing if they jumped to a new minor release.

It also prevents the update screen showing if they downgrade.

Resolves #514