Add a {Name:pdf:ID} Merge tag
In certain circumstances it's useful to just get the Gravity PDF URL which is the purpose of the new {Name:pdf:ID} mergetag. It will function much like the [gravitypdf] shortcode in that it validates a PDF config before displaying the URL. If it isn't valid the merge tag is stripped and an error is placed in the log (Requires the Gravity Forms Logging plugin).

Resolves #404