Fix issue with Advanced Template Manager opening in Safari
Tracked this problem down to Safari's document.activeElement not pointing to the correct dom node when the button was clicked. IE also showed the template manager when enter is pressed on non-ReactJS managed compenents.

I discovered browsers will treat the first <button> tag as the submit if the type isn't explicitly set to "button". Once done, I could remove the activeElement checks and fix the problem.