Swap wp_die() to exit right after outputting PDFA
wp_die() changes the headers we set to correctly display the PDF. While our version of Mpdf has exit right after the PDF generation already, if another plugin ships Mpdf it causes issues.

The way around this is to ensure we correctly exit right after the PDF is output to the browser.