Remove update screen and make the welcome screen unobtrusive
This is to fix issues with caching (both server side and in browsers) which effected
a small percentage of users. The welcome screen should never be intrusive and the
caching issues ensures it WAS instrusive. Because of which, users will now have to
select the "Getting Started" link on the plugin's links.

This has the run-on effect of greatly simplifying the code.

Resolves #609