Create a Global Debug Mode
This removes the Shortcode Debug setting and merges it into a global debug mode.
Also includes template transient caching which is flushed when debug mode is enabled, or when installing / deleting a template.
Finally, PDF debug stats are only shown when Debug mode is enabled and no longer when WP_DEBUG constant is set