Ensure the gfpdf_post_save_pdf action isn't run twice
Move the action so it runs only when the PDF is created and saved to disk. The PDF generation logic already includes the ability to only generate if the PDF doesn't already exist on disk (in the standard flow the PDF is cleaned up after the form submission process, so it's effectively run once).

Resolves #948