Fix error saving PDF settings for the very first time after installation
On the initial installation, the global settings were not being saved correctly due to a loop triggered in the `sanitize_settings` filter. What occurs was `update_option` was called, then our custom sanitize function where we added an update message using `add_settings_error`. This is fine, except when the option doesn't already exist the database. When this is the case `add_option` is called, which subsequently called our `sanitize_settings` filter again. On the second run, because we already added the success message (`add_settings_error`), our transient fallback was triggered.

To prevent this problem, we'll sniff the error variable and see if it's the success message we added. If so, we'll forge ahead.