Resolve cssClass grid repetition using repeater fields
Solves PDF generation problem when using the Nested Forms add-on with Gravity Forms Columns when using a lot of repeater data.

This problems stems from the fact that $form['fields'] contains an array of field objects and our container changes the cssClass value by including our grid classes, when appropriate. This is normally fine, however, when inside a nested forms field those same form fields are processed multiple times and this would inject the grid CSS classes over and over again. This could cause display problems in the PDFs.