Give Nested Forms/Repeater fields unique IDs
The HTML markup generated for fields includes the standard ID "field-#". Inside the Nested Forms/Repeater field, there could be multiples of the same ID. This PR ensures that all IDs are unique for Nested Form fields and Repeater fields.

Because Repeater fields can be nested, to accurately ensure the IDs are always unique a sequential ID was used for each element. Whereas the Nested Forms field can only go one level deep, so the ID used is grouped for each sub-entry.