Always deactivate license key after AJAX request
Regardless of the outcome of the API request, always remove the license key from the database. This prevents UX issues if a license is manually deactivated from the GravityPDF.com account.

Update the error message to let a user know the license may not have been deactivated successfully, with a link to their account.

Resolves #1432