Missing conditional logic rule if no conditional fields in form
When there are no conditional fields in the form, the PDF conditional logic feature is missing the third "rule" input for entry meta rules. This change ensures Gravity Forms correctly displays the rule input first the first rule.

Note: due to the way GFs handles field conditional logic, the same problem occurs if you add a second rule. It's not fixable on our end, but adding a third rule will display the input again. It's such an edge-case this isn't likely to happen often, and the initial rule does work.

Fixes #1510