Resolve conflict with Event Expresso plugin
The EE_Front_Controller::get_request runs on the `parse_request` hook with a priority of 1. For whatever reason, this hook also removes itself when it runs. Under specific circumstances, WordPress will skip hooks on priority 10 due to this removal.

To workaround this bug this PR moves our own `parse_request` hooks to priority 0.